requirements.txt
torch==2.0.0
torchvision==0.15.1
pandas
numpy
scikit-learn
matplotlib
Pillow
